feat: refactors per-tenant options in Finbuckle.MultiTenant.Options package
BREAKING CHANGE: Per-tenant options support was previously part of the Finbuckle.MultiTenant package. Projects will need to reference the Finbuckle.MultiTenant.Options package going forward.
