--- Finbuckle.MultiTenant.sln
@@ -1,87 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-VisualStudioVersion = 15.0.26124.0
-MinimumVisualStudioVersion = 15.0.26124.0
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{CCC3146F-320C-4E1D-8017-056C9DCB6B0C}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Finbuckle.MultiTenant.AspNetCore", "src\Finbuckle.MultiTenant.AspNetCore\Finbuckle.MultiTenant.AspNetCore.csproj", "{CEB62625-8CD8-459B-952F-932A213BD81D}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Finbuckle.MultiTenant.EntityFrameworkCore", "src\Finbuckle.MultiTenant.EntityFrameworkCore\Finbuckle.MultiTenant.EntityFrameworkCore.csproj", "{D71A2E14-3AD0-470F-BED1-E7D4F7874D94}"
-EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "test", "test", "{DDDA7AB6-03C2-4499-BED3-E404CB6ADE69}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Finbuckle.MultiTenant.AspNetCore.Test", "test\Finbuckle.MultiTenant.AspNetCore.Test\Finbuckle.MultiTenant.AspNetCore.Test.csproj", "{CB55C7E5-D1A4-49BC-94D1-9C8CB598426B}"
-EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "samples", "samples", "{78DD95B5-E466-4ED0-8824-36F78AB08174}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Finbuckle.MultiTenant.Test", "test\Finbuckle.MultiTenant.Test\Finbuckle.MultiTenant.Test.csproj", "{4B3A1C30-D42B-43F0-8541-4AB5B51CD05C}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Finbuckle.MultiTenant", "src\Finbuckle.MultiTenant\Finbuckle.MultiTenant.csproj", "{9BA8381A-5E44-4571-9CF1-F4469ECEBB02}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Finbuckle.MultiTenant.EntityFrameworkCore.Test", "test\Finbuckle.MultiTenant.EntityFrameworkCore.Test\Finbuckle.MultiTenant.EntityFrameworkCore.Test.csproj", "{E1A6EBD0-BBB1-4EC3-9318-2D3A55E6868A}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Finbuckle.MultiTenant.Abstractions", "src/Finbuckle.MultiTenant.Abstractions\Finbuckle.MultiTenant.Abstractions.csproj", "{A893BA35-102B-46CD-B4EA-82FEF909D4DF}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Finbuckle.MultiTenant.Identity.EntityFrameworkCore", "src\Finbuckle.MultiTenant.Identity.EntityFrameworkCore\Finbuckle.MultiTenant.Identity.EntityFrameworkCore.csproj", "{F2243E66-209C-490E-99A2-EE71DF5CEEFA}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Finbucke.MultiTenant.Identity.EntityFrameworkCore.Test", "test\Finbucke.MultiTenant.Identity.EntityFrameworkCore.Test\Finbucke.MultiTenant.Identity.EntityFrameworkCore.Test.csproj", "{A673180E-8490-422E-977F-C4976E79C9BD}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{CEB62625-8CD8-459B-952F-932A213BD81D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{CEB62625-8CD8-459B-952F-932A213BD81D}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{CEB62625-8CD8-459B-952F-932A213BD81D}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{CEB62625-8CD8-459B-952F-932A213BD81D}.Release|Any CPU.Build.0 = Release|Any CPU
-		{D71A2E14-3AD0-470F-BED1-E7D4F7874D94}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{D71A2E14-3AD0-470F-BED1-E7D4F7874D94}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{D71A2E14-3AD0-470F-BED1-E7D4F7874D94}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{D71A2E14-3AD0-470F-BED1-E7D4F7874D94}.Release|Any CPU.Build.0 = Release|Any CPU
-		{CB55C7E5-D1A4-49BC-94D1-9C8CB598426B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{CB55C7E5-D1A4-49BC-94D1-9C8CB598426B}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{CB55C7E5-D1A4-49BC-94D1-9C8CB598426B}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{CB55C7E5-D1A4-49BC-94D1-9C8CB598426B}.Release|Any CPU.Build.0 = Release|Any CPU
-		{4B3A1C30-D42B-43F0-8541-4AB5B51CD05C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{4B3A1C30-D42B-43F0-8541-4AB5B51CD05C}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{4B3A1C30-D42B-43F0-8541-4AB5B51CD05C}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{4B3A1C30-D42B-43F0-8541-4AB5B51CD05C}.Release|Any CPU.Build.0 = Release|Any CPU
-		{9BA8381A-5E44-4571-9CF1-F4469ECEBB02}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{9BA8381A-5E44-4571-9CF1-F4469ECEBB02}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{9BA8381A-5E44-4571-9CF1-F4469ECEBB02}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{9BA8381A-5E44-4571-9CF1-F4469ECEBB02}.Release|Any CPU.Build.0 = Release|Any CPU
-		{E1A6EBD0-BBB1-4EC3-9318-2D3A55E6868A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{E1A6EBD0-BBB1-4EC3-9318-2D3A55E6868A}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{E1A6EBD0-BBB1-4EC3-9318-2D3A55E6868A}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{E1A6EBD0-BBB1-4EC3-9318-2D3A55E6868A}.Release|Any CPU.Build.0 = Release|Any CPU
-		{A893BA35-102B-46CD-B4EA-82FEF909D4DF}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{A893BA35-102B-46CD-B4EA-82FEF909D4DF}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{A893BA35-102B-46CD-B4EA-82FEF909D4DF}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{A893BA35-102B-46CD-B4EA-82FEF909D4DF}.Release|Any CPU.Build.0 = Release|Any CPU
-		{F2243E66-209C-490E-99A2-EE71DF5CEEFA}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{F2243E66-209C-490E-99A2-EE71DF5CEEFA}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{F2243E66-209C-490E-99A2-EE71DF5CEEFA}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{F2243E66-209C-490E-99A2-EE71DF5CEEFA}.Release|Any CPU.Build.0 = Release|Any CPU
-		{A673180E-8490-422E-977F-C4976E79C9BD}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{A673180E-8490-422E-977F-C4976E79C9BD}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{A673180E-8490-422E-977F-C4976E79C9BD}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{A673180E-8490-422E-977F-C4976E79C9BD}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(NestedProjects) = preSolution
-		{CEB62625-8CD8-459B-952F-932A213BD81D} = {CCC3146F-320C-4E1D-8017-056C9DCB6B0C}
-		{D71A2E14-3AD0-470F-BED1-E7D4F7874D94} = {CCC3146F-320C-4E1D-8017-056C9DCB6B0C}
-		{CB55C7E5-D1A4-49BC-94D1-9C8CB598426B} = {DDDA7AB6-03C2-4499-BED3-E404CB6ADE69}
-		{4B3A1C30-D42B-43F0-8541-4AB5B51CD05C} = {DDDA7AB6-03C2-4499-BED3-E404CB6ADE69}
-		{9BA8381A-5E44-4571-9CF1-F4469ECEBB02} = {CCC3146F-320C-4E1D-8017-056C9DCB6B0C}
-		{E1A6EBD0-BBB1-4EC3-9318-2D3A55E6868A} = {DDDA7AB6-03C2-4499-BED3-E404CB6ADE69}
-		{A893BA35-102B-46CD-B4EA-82FEF909D4DF} = {CCC3146F-320C-4E1D-8017-056C9DCB6B0C}
-		{F2243E66-209C-490E-99A2-EE71DF5CEEFA} = {CCC3146F-320C-4E1D-8017-056C9DCB6B0C}
-		{A673180E-8490-422E-977F-C4976E79C9BD} = {DDDA7AB6-03C2-4499-BED3-E404CB6ADE69}
-	EndGlobalSection
-EndGlobal
