+++ Finbuckle.MultiTenant.slnx
@@ -1,0 +1,18 @@
+﻿<Solution>
+  <Folder Name="/samples/" />
+  <Folder Name="/src/">
+    <Project Path="src/Finbuckle.MultiTenant.Abstractions\Finbuckle.MultiTenant.Abstractions.csproj" Type="Classic C#" />
+    <Project Path="src\Finbuckle.MultiTenant.AspNetCore\Finbuckle.MultiTenant.AspNetCore.csproj" Type="Classic C#" />
+    <Project Path="src\Finbuckle.MultiTenant.EntityFrameworkCore\Finbuckle.MultiTenant.EntityFrameworkCore.csproj" Type="Classic C#" />
+    <Project Path="src\Finbuckle.MultiTenant.Identity.EntityFrameworkCore\Finbuckle.MultiTenant.Identity.EntityFrameworkCore.csproj" Type="Classic C#" />
+    <Project Path="src\Finbuckle.MultiTenant.Options\Finbuckle.MultiTenant.Options.csproj" Type="Classic C#" />
+    <Project Path="src\Finbuckle.MultiTenant\Finbuckle.MultiTenant.csproj" Type="Classic C#" />
+  </Folder>
+  <Folder Name="/test/">
+    <Project Path="test\Finbucke.MultiTenant.Identity.EntityFrameworkCore.Test\Finbucke.MultiTenant.Identity.EntityFrameworkCore.Test.csproj" Type="Classic C#" />
+    <Project Path="test\Finbuckle.MultiTenant.AspNetCore.Test\Finbuckle.MultiTenant.AspNetCore.Test.csproj" Type="Classic C#" />
+    <Project Path="test\Finbuckle.MultiTenant.EntityFrameworkCore.Test\Finbuckle.MultiTenant.EntityFrameworkCore.Test.csproj" Type="Classic C#" />
+    <Project Path="test\Finbuckle.MultiTenant.Options.Test\Finbuckle.MultiTenant.Options.Test.csproj" Type="Classic C#" />
+    <Project Path="test\Finbuckle.MultiTenant.Test\Finbuckle.MultiTenant.Test.csproj" Type="Classic C#" />
+  </Folder>
+</Solution>
