+++ src/Finbuckle.MultiTenant.Abstractions/AssemblyInfo.cs
@@ -9,3 +9,4 @@ using System.Runtime.CompilerServices;
 [assembly:InternalsVisibleTo("Finbuckle.MultiTenant.AspNetCore.Test")]
 [assembly:InternalsVisibleTo("Finbuckle.MultiTenant.EntityFrameworkCore")]
 [assembly:InternalsVisibleTo("Finbuckle.MultiTenant.EntityFrameworkCore.Test")]
+[assembly:InternalsVisibleTo("Finbuckle.MultiTenant.Options.Test")]
