+++ src/Finbuckle.MultiTenant.AspNetCore/Finbuckle.MultiTenant.AspNetCore.csproj
@@ -10,5 +10,6 @@
         <PackageReference Include="Microsoft.AspNetCore.Authentication.OpenIdConnect" Version="10.0.0-*"/>
        
         <ProjectReference Include="../Finbuckle.MultiTenant/Finbuckle.MultiTenant.csproj"/>
+        <ProjectReference Include="../Finbuckle.MultiTenant.Options/Finbuckle.MultiTenant.Options.csproj"/>
     </ItemGroup>
 </Project>
