+++ src/Finbuckle.MultiTenant.EntityFrameworkCore/Extensions/MultiTenantDbContextExtensions.cs
@@ -1,11 +1,9 @@
 // Copyright Finbuckle LLC, Andrew White, and Contributors.
 // Refer to the solution LICENSE file for more information.
 
-using System.Net.Mime;
 using Finbuckle.MultiTenant.Abstractions;
 using Finbuckle.MultiTenant.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore;
-using Microsoft.EntityFrameworkCore.ChangeTracking;
 
 // ReSharper disable once CheckNamespace
 namespace Finbuckle.MultiTenant;
