+++ src/Finbuckle.MultiTenant.EntityFrameworkCore/MultiTenantEntityTypeBuilder.cs
@@ -2,7 +2,6 @@
 // Refer to the solution LICENSE file for more information.
 
 using System.Collections.Immutable;
-using System.Diagnostics.CodeAnalysis;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Metadata;
 using Microsoft.EntityFrameworkCore.Metadata.Builders;
