+++ src/Finbuckle.MultiTenant.Options/Extensions/ServiceCollectionExtensions.cs
@@ -1,0 +1,170 @@
+// Copyright Finbuckle LLC, Andrew White, and Contributors.
+// Refer to the solution LICENSE file for more information.
+
+using Finbuckle.MultiTenant.Abstractions;
+using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.DependencyInjection.Extensions;
+using Microsoft.Extensions.Options;
+
+namespace Finbuckle.MultiTenant.Options.Extensions;
+
+/// <summary>
+/// IServiceCollection extension methods for Finbuckle.MultiTenant.Options.
+/// </summary>
+public static class FinbuckleServiceCollectionExtensions
+{
+    /// <summary>
+    /// Registers an action used to configure a particular type of options per tenant.
+    /// </summary>
+    /// <typeparam name="TOptions">The options type to be configured.</typeparam>
+    /// <typeparam name="TTenantInfo">The ITenantInfo implementation type.</typeparam>
+    /// <param name="services">The <see cref="IServiceCollection"/> to add the services to.</param>
+    /// <param name="name">The name of the options instance.</param>
+    /// <param name="configureOptions">The action used to configure the options.</param>
+    /// <returns>The <see cref="IServiceCollection"/> so that additional calls can be chained.</returns>
+    public static IServiceCollection ConfigurePerTenant<TOptions, TTenantInfo>(
+        this IServiceCollection services,
+        string? name, Action<TOptions, TTenantInfo> configureOptions)
+        where TOptions : class
+        where TTenantInfo : class, ITenantInfo, new()
+    {
+        ConfigurePerTenantReqs<TOptions>(services);
+
+        services.AddTransient<IConfigureOptions<TOptions>>(sp =>
+            new ConfigureNamedOptions<TOptions, IMultiTenantContextAccessor<TTenantInfo>>(
+                name,
+                sp.GetRequiredService<IMultiTenantContextAccessor<TTenantInfo>>(),
+                (options, mtcAccessor) =>
+                {
+                    var tenantInfo = mtcAccessor.MultiTenantContext?.TenantInfo;
+                    if (tenantInfo is not null)
+                        configureOptions(options, tenantInfo);
+                }));
+
+        return services;
+    }
+
+    /// <summary>
+    /// Registers an action used to configure a particular type of options.
+    /// </summary>
+    /// <typeparam name="TOptions">The options type to be configured.</typeparam>
+    /// <typeparam name="TTenantInfo">The ITenantInfo implementation type.</typeparam>
+    /// <param name="services">The <see cref="IServiceCollection"/> to add the services to.</param>
+    /// <param name="configureOptions">The action used to configure the options.</param>
+    /// <returns>The <see cref="IServiceCollection"/> so that additional calls can be chained.</returns>
+    public static IServiceCollection ConfigurePerTenant<TOptions, TTenantInfo>(
+        this IServiceCollection services,
+        Action<TOptions, TTenantInfo> configureOptions)
+        where TOptions : class
+        where TTenantInfo : class, ITenantInfo, new()
+    {
+        return services.ConfigurePerTenant(Microsoft.Extensions.Options.Options.DefaultName, configureOptions);
+    }
+
+    /// <summary>
+    /// Registers an action used to configure all instances of a particular type of options per tenant.
+    /// </summary>
+    /// <typeparam name="TOptions">The options type to be configured.</typeparam>
+    /// <typeparam name="TTenantInfo">The ITenantInfo implementation type.</typeparam>
+    /// <param name="services">The <see cref="IServiceCollection"/> to add the services to.</param>
+    /// <param name="configureOptions">The action used to configure the options.</param>
+    /// <returns>The <see cref="IServiceCollection"/> so that additional calls can be chained.</returns>
+    public static IServiceCollection ConfigureAllPerTenant<TOptions, TTenantInfo>(
+        this IServiceCollection services,
+        Action<TOptions, TTenantInfo> configureOptions)
+        where TOptions : class
+        where TTenantInfo : class, ITenantInfo, new()
+    {
+        return services.ConfigurePerTenant(null, configureOptions);
+    }
+
+    /// <summary>
+    /// Registers a post configure action used to configure a particular type of options per tenant.
+    /// </summary>
+    /// <typeparam name="TOptions">The options type to be configured.</typeparam>
+    /// <typeparam name="TTenantInfo">The ITenantInfo implementation type.</typeparam>
+    /// <param name="services">The <see cref="IServiceCollection"/> to add the services to.</param>
+    /// <param name="name">The name of the options instance.</param>
+    /// <param name="configureOptions">The action used to configure the options.</param>
+    /// <returns>The <see cref="IServiceCollection"/> so that additional calls can be chained.</returns>
+    public static IServiceCollection PostConfigurePerTenant<TOptions, TTenantInfo>(
+        this IServiceCollection services,
+        string? name, Action<TOptions, TTenantInfo> configureOptions)
+        where TOptions : class
+        where TTenantInfo : class, ITenantInfo, new()
+    {
+        ConfigurePerTenantReqs<TOptions>(services);
+
+        services.AddTransient<IPostConfigureOptions<TOptions>>(sp =>
+            new PostConfigureOptions<TOptions, IMultiTenantContextAccessor<TTenantInfo>>(
+                name,
+                sp.GetRequiredService<IMultiTenantContextAccessor<TTenantInfo>>(),
+                (options, mtcAccessor) =>
+                {
+                    var tenantInfo = mtcAccessor.MultiTenantContext?.TenantInfo;
+                    if (tenantInfo is not null)
+                        configureOptions(options, tenantInfo);
+                }));
+
+        return services;
+    }
+
+    /// <summary>
+    /// Registers a post configure action used to configure a particular type of options per tenant.
+    /// </summary>
+    /// <typeparam name="TOptions">The options type to be configured.</typeparam>
+    /// <typeparam name="TTenantInfo">The ITenantInfo implementation type.</typeparam>
+    /// <param name="services">The <see cref="IServiceCollection"/> to add the services to.</param>
+    /// <param name="configureOptions">The action used to configure the options.</param>
+    /// <returns>The <see cref="IServiceCollection"/> so that additional calls can be chained.</returns>
+    public static IServiceCollection PostConfigurePerTenant<TOptions, TTenantInfo>(
+        this IServiceCollection services,
+        Action<TOptions, TTenantInfo> configureOptions)
+        where TOptions : class
+        where TTenantInfo : class, ITenantInfo, new()
+    {
+        return services.PostConfigurePerTenant(Microsoft.Extensions.Options.Options.DefaultName, configureOptions);
+    }
+
+    /// <summary>
+    /// Registers a post configure action used to configure all instances of a particular type of options per tenant.
+    /// </summary>
+    /// <typeparam name="TOptions">The options type to be configured.</typeparam>
+    /// <typeparam name="TTenantInfo">The ITenantInfo implementation type.</typeparam>
+    /// <param name="services">The <see cref="IServiceCollection"/> to add the services to.</param>
+    /// <param name="configureOptions">The action used to configure the options.</param>
+    /// <returns>The <see cref="IServiceCollection"/> so that additional calls can be chained.</returns>
+    public static IServiceCollection PostConfigureAllPerTenant<TOptions, TTenantInfo>(
+        this IServiceCollection services,
+        Action<TOptions, TTenantInfo> configureOptions)
+        where TOptions : class
+        where TTenantInfo : class, ITenantInfo, new()
+    {
+        return services.PostConfigurePerTenant(null, configureOptions);
+    }
+
+    /// <summary>
+    /// Configures the required services for per-tenant options support.
+    /// </summary>
+    /// <typeparam name="TOptions">The options type to be configured.</typeparam>
+    /// <param name="services">The <see cref="IServiceCollection"/> to add the services to.</param>
+    internal static void ConfigurePerTenantReqs<TOptions>(IServiceCollection services)
+        where TOptions : class
+    {
+        ArgumentNullException.ThrowIfNull(services);
+
+        // Required infrastructure.
+        services.AddOptions();
+
+        services.AddSingleton<IOptionsMonitorCache<TOptions>, MultiTenantOptionsCache<TOptions>>();
+        services.AddScoped<IOptionsSnapshot<TOptions>>(BuildOptionsManager);
+        services.AddSingleton<IOptions<TOptions>>(BuildOptionsManager);
+        return;
+
+        MultiTenantOptionsManager<TOptions> BuildOptionsManager(IServiceProvider sp)
+        {
+            IOptionsMonitorCache<TOptions> cache = ActivatorUtilities.CreateInstance<MultiTenantOptionsCache<TOptions>>(sp);
+            return ActivatorUtilities.CreateInstance<MultiTenantOptionsManager<TOptions>>(sp, cache);
+        }
+    }
+}
