+++ src/Finbuckle.MultiTenant.Options/Finbuckle.MultiTenant.Options.csproj
@@ -1,0 +1,16 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+    <PropertyGroup>
+        <TargetFramework>net10.0</TargetFramework>
+        <Title>Finbuckle.MultiTenant.Options</Title>
+        <Description>Per-Tenant library package for Finbuckle.MultiTenant.</Description>
+    </PropertyGroup>
+
+    <ItemGroup>
+        <PackageReference Include="Microsoft.Extensions.Logging" Version="10.0.0-*"/>
+        <PackageReference Include="Microsoft.Extensions.Options" Version="10.0.0-*"/>
+        
+        <ProjectReference Include="../Finbuckle.MultiTenant.Abstractions/Finbuckle.MultiTenant.Abstractions.csproj"/>
+    </ItemGroup>
+
+</Project>
