+++ src/Finbuckle.MultiTenant/DependencyInjection/ServiceCollectionExtensions.cs
@@ -3,10 +3,7 @@
 
 using System.Diagnostics.CodeAnalysis;
 using Finbuckle.MultiTenant.Abstractions;
-using Finbuckle.MultiTenant.Options;
 using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.DependencyInjection.Extensions;
-using Microsoft.Extensions.Options;
 
 // ReSharper disable once CheckNamespace
 namespace Finbuckle.MultiTenant;
@@ -75,10 +72,9 @@ public static class FinbuckleServiceCollectionExtensions
         if (existingServices.Count == 0)
             throw new ArgumentException($"No service of type {typeof(TService).Name} found.");
 
-        ServiceDescriptor? newService;
-
         foreach (var existingService in existingServices)
         {
+            ServiceDescriptor? newService;
             if (existingService.ImplementationType is not null)
             {
                 newService = new ServiceDescriptor(existingService.ServiceType,
@@ -146,161 +142,5 @@ public static class FinbuckleServiceCollectionExtensions
         }
 
         return true;
-    }
-
-    /// <summary>
-    /// Registers an action used to configure a particular type of options per tenant.
-    /// </summary>
-    /// <typeparam name="TOptions">The options type to be configured.</typeparam>
-    /// <typeparam name="TTenantInfo">The ITenantInfo implementation type.</typeparam>
-    /// <param name="services">The <see cref="IServiceCollection"/> to add the services to.</param>
-    /// <param name="name">The name of the options instance.</param>
-    /// <param name="configureOptions">The action used to configure the options.</param>
-    /// <returns>The <see cref="IServiceCollection"/> so that additional calls can be chained.</returns>
-    public static IServiceCollection ConfigurePerTenant<TOptions, TTenantInfo>(
-        this IServiceCollection services,
-        string? name, Action<TOptions, TTenantInfo> configureOptions)
-        where TOptions : class
-        where TTenantInfo : class, ITenantInfo, new()
-    {
-        ConfigurePerTenantReqs<TOptions>(services);
-
-        services.AddTransient<IConfigureOptions<TOptions>>(sp =>
-            new ConfigureNamedOptions<TOptions, IMultiTenantContextAccessor<TTenantInfo>>(
-                name,
-                sp.GetRequiredService<IMultiTenantContextAccessor<TTenantInfo>>(),
-                (options, mtcAccessor) =>
-                {
-                    var tenantInfo = mtcAccessor.MultiTenantContext?.TenantInfo;
-                    if (tenantInfo is not null)
-                        configureOptions(options, tenantInfo);
-                }));
-
-        return services;
-    }
-
-    /// <summary>
-    /// Registers an action used to configure a particular type of options.
-    /// </summary>
-    /// <typeparam name="TOptions">The options type to be configured.</typeparam>
-    /// <typeparam name="TTenantInfo">The ITenantInfo implementation type.</typeparam>
-    /// <param name="services">The <see cref="IServiceCollection"/> to add the services to.</param>
-    /// <param name="configureOptions">The action used to configure the options.</param>
-    /// <returns>The <see cref="IServiceCollection"/> so that additional calls can be chained.</returns>
-    public static IServiceCollection ConfigurePerTenant<TOptions, TTenantInfo>(
-        this IServiceCollection services,
-        Action<TOptions, TTenantInfo> configureOptions)
-        where TOptions : class
-        where TTenantInfo : class, ITenantInfo, new()
-    {
-        return services.ConfigurePerTenant(Microsoft.Extensions.Options.Options.DefaultName, configureOptions);
-    }
-
-    /// <summary>
-    /// Registers an action used to configure all instances of a particular type of options per tenant.
-    /// </summary>
-    /// <typeparam name="TOptions">The options type to be configured.</typeparam>
-    /// <typeparam name="TTenantInfo">The ITenantInfo implementation type.</typeparam>
-    /// <param name="services">The <see cref="IServiceCollection"/> to add the services to.</param>
-    /// <param name="configureOptions">The action used to configure the options.</param>
-    /// <returns>The <see cref="IServiceCollection"/> so that additional calls can be chained.</returns>
-    public static IServiceCollection ConfigureAllPerTenant<TOptions, TTenantInfo>(
-        this IServiceCollection services,
-        Action<TOptions, TTenantInfo> configureOptions)
-        where TOptions : class
-        where TTenantInfo : class, ITenantInfo, new()
-    {
-        return services.ConfigurePerTenant(null, configureOptions);
-    }
-
-    /// <summary>
-    /// Registers a post configure action used to configure a particular type of options per tenant.
-    /// </summary>
-    /// <typeparam name="TOptions">The options type to be configured.</typeparam>
-    /// <typeparam name="TTenantInfo">The ITenantInfo implementation type.</typeparam>
-    /// <param name="services">The <see cref="IServiceCollection"/> to add the services to.</param>
-    /// <param name="name">The name of the options instance.</param>
-    /// <param name="configureOptions">The action used to configure the options.</param>
-    /// <returns>The <see cref="IServiceCollection"/> so that additional calls can be chained.</returns>
-    public static IServiceCollection PostConfigurePerTenant<TOptions, TTenantInfo>(
-        this IServiceCollection services,
-        string? name, Action<TOptions, TTenantInfo> configureOptions)
-        where TOptions : class
-        where TTenantInfo : class, ITenantInfo, new()
-    {
-        ConfigurePerTenantReqs<TOptions>(services);
-
-        services.AddTransient<IPostConfigureOptions<TOptions>>(sp =>
-            new PostConfigureOptions<TOptions, IMultiTenantContextAccessor<TTenantInfo>>(
-                name,
-                sp.GetRequiredService<IMultiTenantContextAccessor<TTenantInfo>>(),
-                (options, mtcAccessor) =>
-                {
-                    var tenantInfo = mtcAccessor.MultiTenantContext?.TenantInfo;
-                    if (tenantInfo is not null)
-                        configureOptions(options, tenantInfo);
-                }));
-
-        return services;
-    }
-
-    /// <summary>
-    /// Registers a post configure action used to configure a particular type of options per tenant.
-    /// </summary>
-    /// <typeparam name="TOptions">The options type to be configured.</typeparam>
-    /// <typeparam name="TTenantInfo">The ITenantInfo implementation type.</typeparam>
-    /// <param name="services">The <see cref="IServiceCollection"/> to add the services to.</param>
-    /// <param name="configureOptions">The action used to configure the options.</param>
-    /// <returns>The <see cref="IServiceCollection"/> so that additional calls can be chained.</returns>
-    public static IServiceCollection PostConfigurePerTenant<TOptions, TTenantInfo>(
-        this IServiceCollection services,
-        Action<TOptions, TTenantInfo> configureOptions)
-        where TOptions : class
-        where TTenantInfo : class, ITenantInfo, new()
-    {
-        return services.PostConfigurePerTenant(Microsoft.Extensions.Options.Options.DefaultName, configureOptions);
-    }
-
-    /// <summary>
-    /// Registers a post configure action used to configure all instances of a particular type of options per tenant.
-    /// </summary>
-    /// <typeparam name="TOptions">The options type to be configured.</typeparam>
-    /// <typeparam name="TTenantInfo">The ITenantInfo implementation type.</typeparam>
-    /// <param name="services">The <see cref="IServiceCollection"/> to add the services to.</param>
-    /// <param name="configureOptions">The action used to configure the options.</param>
-    /// <returns>The <see cref="IServiceCollection"/> so that additional calls can be chained.</returns>
-    public static IServiceCollection PostConfigureAllPerTenant<TOptions, TTenantInfo>(
-        this IServiceCollection services,
-        Action<TOptions, TTenantInfo> configureOptions)
-        where TOptions : class
-        where TTenantInfo : class, ITenantInfo, new()
-    {
-        return services.PostConfigurePerTenant(null, configureOptions);
-    }
-
-    /// <summary>
-    /// Configures the required services for per-tenant options support.
-    /// </summary>
-    /// <typeparam name="TOptions">The options type to be configured.</typeparam>
-    /// <param name="services">The <see cref="IServiceCollection"/> to add the services to.</param>
-    internal static void ConfigurePerTenantReqs<TOptions>(IServiceCollection services)
-        where TOptions : class
-    {
-        ArgumentNullException.ThrowIfNull(services);
-
-        // Required infrastructure.
-        services.AddOptions();
-
-        // TODO: Add check for success
-        services.TryAddSingleton<IOptionsMonitorCache<TOptions>, MultiTenantOptionsCache<TOptions>>();
-        services.TryAddScoped<IOptionsSnapshot<TOptions>>(BuildOptionsManager);
-        services.TryAddSingleton<IOptions<TOptions>>(BuildOptionsManager);
-        return;
-
-        MultiTenantOptionsManager<TOptions> BuildOptionsManager(IServiceProvider sp)
-        {
-            IOptionsMonitorCache<TOptions> cache = ActivatorUtilities.CreateInstance<MultiTenantOptionsCache<TOptions>>(sp);
-            return ActivatorUtilities.CreateInstance<MultiTenantOptionsManager<TOptions>>(sp, cache);
-        }
     }
 }
