+++ src/Finbuckle.MultiTenant/Finbuckle.MultiTenant.csproj
@@ -8,7 +8,7 @@
     <ItemGroup>
         <PackageReference Include="Microsoft.Extensions.Logging" Version="10.0.0-*"/>
         <PackageReference Include="Microsoft.Extensions.DependencyInjection" Version="10.0.0-*"/>
-        <PackageReference Include="Microsoft.Extensions.Options" Version="10.0.0-*"/>
+        
         <PackageReference Include="Microsoft.Extensions.Options.ConfigurationExtensions" Version="10.0.0-*"/>
         <PackageReference Include="Microsoft.Extensions.Configuration" Version="10.0.0-*"/>
         <PackageReference Include="Microsoft.Extensions.Http" Version="10.0.0-*"/>
