+++ test/Finbuckle.MultiTenant.Options.Test/MultiTenantOptionsCacheShould.cs
@@ -5,11 +5,10 @@ using System.Collections.Concurrent;
 using System.ComponentModel.DataAnnotations;
 using System.Reflection;
 using Finbuckle.MultiTenant.Abstractions;
-using Finbuckle.MultiTenant.Options;
 using Microsoft.Extensions.Options;
 using Xunit;
 
-namespace Finbuckle.MultiTenant.Test.Options;
+namespace Finbuckle.MultiTenant.Options.Test;
 
 public class MultiTenantOptionsCacheShould
 {
