+++ test/Finbuckle.MultiTenant.Test/DependencyInjection/ServiceCollectionExtensionsShould.cs
@@ -13,119 +13,68 @@ public class ServiceCollectionExtensionsShould
     [Fact]
     public void RegisterITenantResolverInDi()
     {
-            var services = new ServiceCollection();
-            services.AddMultiTenant<TenantInfo>();
+        var services = new ServiceCollection();
+        services.AddMultiTenant<TenantInfo>();
 
-            var service = services.SingleOrDefault(s => s.ServiceType == typeof(ITenantResolver));
+        var service = services.SingleOrDefault(s => s.ServiceType == typeof(ITenantResolver));
 
-            Assert.NotNull(service);
-            Assert.Equal(ServiceLifetime.Scoped, service.Lifetime);
-        }
+        Assert.NotNull(service);
+        Assert.Equal(ServiceLifetime.Scoped, service.Lifetime);
+    }
 
     [Fact]
     public void RegisterITenantResolverGenericInDi()
     {
-            var services = new ServiceCollection();
-            services.AddMultiTenant<TenantInfo>();
+        var services = new ServiceCollection();
+        services.AddMultiTenant<TenantInfo>();
 
-            var service = services.SingleOrDefault(s => s.ServiceType == typeof(ITenantResolver<TenantInfo>));
+        var service = services.SingleOrDefault(s => s.ServiceType == typeof(ITenantResolver<TenantInfo>));
 
-            Assert.NotNull(service);
-            Assert.Equal(ServiceLifetime.Scoped, service.Lifetime);
-        }
+        Assert.NotNull(service);
+        Assert.Equal(ServiceLifetime.Scoped, service.Lifetime);
+    }
 
     [Fact]
     public void RegisterIMultiTenantContextAccessorInDi()
     {
-            var services = new ServiceCollection();
-            services.AddMultiTenant<TenantInfo>();
+        var services = new ServiceCollection();
+        services.AddMultiTenant<TenantInfo>();
 
-            var service = services.SingleOrDefault(s => s.Lifetime == ServiceLifetime.Singleton &&
-                                                        s.ServiceType == typeof(IMultiTenantContextAccessor));
+        var service = services.SingleOrDefault(s => s.Lifetime == ServiceLifetime.Singleton &&
+                                                    s.ServiceType == typeof(IMultiTenantContextAccessor));
 
-            Assert.NotNull(service);
-            Assert.Equal(ServiceLifetime.Singleton, service.Lifetime);
-        }
+        Assert.NotNull(service);
+        Assert.Equal(ServiceLifetime.Singleton, service.Lifetime);
+    }
 
     [Fact]
     public void RegisterIMultiTenantContextAccessorGenericInDi()
     {
-            var services = new ServiceCollection();
-            services.AddMultiTenant<TenantInfo>();
+        var services = new ServiceCollection();
+        services.AddMultiTenant<TenantInfo>();
 
-            var service = services.SingleOrDefault(s => s.Lifetime == ServiceLifetime.Singleton &&
-                                                        s.ServiceType ==
-                                                        typeof(IMultiTenantContextAccessor<TenantInfo>));
+        var service = services.SingleOrDefault(s => s.Lifetime == ServiceLifetime.Singleton &&
+                                                    s.ServiceType ==
+                                                    typeof(IMultiTenantContextAccessor<TenantInfo>));
 
-            Assert.NotNull(service);
-            Assert.Equal(ServiceLifetime.Singleton, service.Lifetime);
-        }
+        Assert.NotNull(service);
+        Assert.Equal(ServiceLifetime.Singleton, service.Lifetime);
+    }
 
     [Fact]
     public void RegisterMultiTenantOptionsInDi()
     {
-            var services = new ServiceCollection();
-            services.AddMultiTenant<TenantInfo>();
+        var services = new ServiceCollection();
+        services.AddMultiTenant<TenantInfo>();
 
-            var service = services.FirstOrDefault(s => s.Lifetime == ServiceLifetime.Singleton &&
-                                                        s.ServiceType == typeof(IConfigureOptions<MultiTenantOptions<TenantInfo>>));
+        var service = services.FirstOrDefault(s => s.Lifetime == ServiceLifetime.Singleton &&
+                                                   s.ServiceType ==
+                                                   typeof(IConfigureOptions<MultiTenantOptions<TenantInfo>>));
 
-            Assert.NotNull(service);
-            Assert.Equal(ServiceLifetime.Singleton, service.Lifetime);
-        }
-
-    public class TestOptions
-    {
-        public string? Prop1 { get; set; }
+        Assert.NotNull(service);
+        Assert.Equal(ServiceLifetime.Singleton, service.Lifetime);
     }
 
-    [Fact]
-    public void RegisterNamedOptionsPerTenant()
-    {
-            var services = new ServiceCollection();
-            services.AddMultiTenant<TenantInfo>();
-            services.ConfigurePerTenant<TestOptions, TenantInfo>("name1",
-                (option, tenant) => option.Prop1 = tenant.Id);
-            var sp = services.BuildServiceProvider();
-
-            var configs = sp.GetRequiredService<IEnumerable<IConfigureOptions<TestOptions>>>();
-            var config = configs.Where(config => config is ConfigureNamedOptions<TestOptions, IMultiTenantContextAccessor<TenantInfo>> options).ToList();
-
-            Assert.Single(config);
-            Assert.Equal("name1", config.Select(c => (ConfigureNamedOptions<TestOptions, IMultiTenantContextAccessor<TenantInfo>>)c).Single().Name);
-        }
-
-    [Fact]
-    public void RegisterUnnamedOptionsPerTenant()
-    {
-            var services = new ServiceCollection();
-            services.AddMultiTenant<TenantInfo>();
-            services.ConfigurePerTenant<TestOptions, TenantInfo>((option, tenant) => option.Prop1 = tenant.Id);
-            var sp = services.BuildServiceProvider();
-
-            var configs = sp.GetRequiredService<IEnumerable<IConfigureOptions<TestOptions>>>();
-            var config = configs.Where(config => config is ConfigureNamedOptions<TestOptions, IMultiTenantContextAccessor<TenantInfo>> options).ToList();
-
-            Assert.Single(config);
-            Assert.Equal(Microsoft.Extensions.Options.Options.DefaultName,
-                config.Select(c => (ConfigureNamedOptions<TestOptions, IMultiTenantContextAccessor<TenantInfo>>)c).Single().Name);
-        }
-        
-    [Fact]
-    public void RegisterAllOptionsPerTenant()
-    {
-            var services = new ServiceCollection();
-            services.AddMultiTenant<TenantInfo>();
-            services.ConfigureAllPerTenant<TestOptions, TenantInfo>((option, tenant) => option.Prop1 = tenant.Id);
-            var sp = services.BuildServiceProvider();
-
-            var configs = sp.GetRequiredService<IEnumerable<IConfigureOptions<TestOptions>>>();
-            var config = configs.Where(config => config is ConfigureNamedOptions<TestOptions, IMultiTenantContextAccessor<TenantInfo>> options).ToList();
-
-            Assert.Single(config);
-            Assert.Null(config.Select(c => (ConfigureNamedOptions<TestOptions, IMultiTenantContextAccessor<TenantInfo>>)c).Single().Name);
-        }
-        
     [Fact]
     public void DecorateService_WithImplementationType_WrapsService()
     {
@@ -192,8 +141,14 @@ public class ServiceCollectionExtensionsShould
         Assert.Contains(instance2, inners);
     }
 
-    public interface ITestService { }
-    public class TestService : ITestService { }
+    public interface ITestService
+    {
+    }
+
+    public class TestService : ITestService
+    {
+    }
+
     public class DecoratedTestService : ITestService
     {
         public ITestService Inner { get; }
